Adds merge ready check (pending actual versions). Clean up and follow gqlgen recommended practices
