+++ gqlgen.yml
@@ -37,8 +37,8 @@ resolver:
 
 # gqlgen will search for any type names in the schema in these go packages
 # if they match it will use them, otherwise it will generate them.
-autobind:
-  - "eth2-crawler/graph/model"
+#autobind:
+#  - "eth2-crawler/graph/model"
 
 # This section declares type mapping between the GraphQL and go type systems
 #
