+++ graph/schema.graphqls
@@ -49,4 +49,5 @@ type Query {
   getNodeStatsOverTime(start: Float!, end: Float!): [NodeStatsOverTime!]!
   getRegionalStats: RegionalStats!
   getAltairUpgradePercentage: Float!
+  getMergeReadyPercentage: Float!
 }
