+++ graph/schema.resolvers.go
@@ -1,6 +1,3 @@
-// Copyright 2021 ChainSafe Systems
-// SPDX-License-Identifier: LGPL-3.0-only
-
 package graph
 
 // This file will be automatically regenerated based on the schema, any resolver implementations
@@ -11,8 +8,6 @@ import (
 	"eth2-crawler/graph/generated"
 	"eth2-crawler/graph/model"
 	svcModels "eth2-crawler/models"
-
-	"github.com/hashicorp/go-version"
 )
 
 func (r *queryResolver) AggregateByAgentName(ctx context.Context) ([]*model.AggregateData, error) {
@@ -209,48 +204,27 @@ func (r *queryResolver) GetAltairUpgradePercentage(ctx context.Context) (float64
 	return percentage, nil
 }
 
+func (r *queryResolver) GetMergeReadyPercentage(ctx context.Context) (float64, error) {
+	aggregateData, err := r.peerStore.AggregateByClientVersion(ctx)
+	if err != nil {
+		return 0, err
+	}
+
+	count := 0
+	total := 0
+	for _, client := range aggregateData {
+		for _, v := range client.Versions {
+			total += v.Count
+			if isMergeReady(client.Client, v.Name) {
+				count += v.Count
+			}
+		}
+	}
+	percentage := float64(count) / float64(total) * 100
+	return percentage, nil
+}
+
 // Query returns generated.QueryResolver implementation.
 func (r *Resolver) Query() generated.QueryResolver { return &queryResolver{r} }
 
 type queryResolver struct{ *Resolver }
-
-func supportAltairUpgrade(clientName, ver string) bool {
-	if len(ver) != 0 && ver[0:1] != "v" {
-		ver = "v" + ver
-	}
-	clientVersion, err := version.NewVersion(ver)
-	if err != nil {
-		return false
-	}
-
-	switch svcModels.ClientName(clientName) {
-	case svcModels.PrysmClient:
-		v, _ := version.NewVersion("v2.0.0")
-		if clientVersion.GreaterThanOrEqual(v) {
-			return true
-		}
-	case svcModels.TekuClient:
-		v, _ := version.NewVersion("v21.9.2")
-		if clientVersion.GreaterThanOrEqual(v) {
-			return true
-		}
-	case svcModels.LighthouseClient:
-		v, _ := version.NewVersion("v2.0.0")
-		if clientVersion.GreaterThanOrEqual(v) {
-			return true
-		}
-	case svcModels.NimbusClient:
-		v, _ := version.NewVersion("v1.5.0")
-		if clientVersion.GreaterThanOrEqual(v) {
-			return true
-		}
-	case svcModels.LodestarClient:
-		v, _ := version.NewVersion("v0.31.0")
-		if clientVersion.GreaterThanOrEqual(v) {
-			return true
-		}
-	default:
-		return false
-	}
-	return false
-}
